Emend citation for term "napkin"
Such a table can be found here: http://conwaylife.com/wiki/Isotropic_non-totalistic_Life-like_cellular_automaton
The original term, "bowling napkin", was contrived from two things in succession. The first is the shape of such a transition table: the interior columns are longer, giving it a triangular shape reminiscent of a set of bowling pins -- and the second is the resemblance of a set of bowling pins' triangular shape to a "napkin or bandana", specifically a triangular napkin tied around the neck. I have no idea how these leaps could possibly have been made to arrive at "bowling napkin", but that's its full etymology!